Update documentation for TLSF and LLFF heaps
