+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- Added a Two-Level Segregated Fit heap with the `tlsf` feature.
+
+### Changed
+
+- The `Heap` struct has been renamed to `LlffHeap` and requires the `llff` feature.
+
 ## [v0.5.1] - 2023-11-04
 
 ### Added
