+++ Cargo.toml
@@ -38,7 +38,6 @@ linked_list_allocator = { version = "0.10.5", default-features = false, optional
 rlsf = { version = "0.2.1", default-features = false, optional = true }
 const-default = { version = "1.0.0", default-features = false, optional = true }
 
-
 [dev-dependencies]
 cortex-m = { version = "0.7.6", features = ["critical-section-single-core"] }
 cortex-m-rt = "0.7"
